Fixed categorical test
